+++ crates/objc2/src/macros.rs
@@ -4,6 +4,7 @@ mod __rewrite_self_arg;
 mod declare_class;
 mod extern_class;
 mod extern_methods;
+mod extern_protocol;
 
 /// Gets a reference to a [`Class`] from the given name.
 ///